Batch monitoring presentation
Somebody asked me to explain a batch job, this is meant to be a general intro of how batch job monitoring and management is done in a prod environment. 

Slide 16 has a flowchart error that I will correct one day.

Chances of me expanding this are practically null, I just dont have time.